refactor(frontend): move ccn 0044 to a dedicated folder
