+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ccn/0044/0044_indemnite.js
@@ -1,4 +1,4 @@
-import { round } from "../math";
+import { round } from "../../math";
 /**
  * calcul de l'indemnite de licenciement conventionnelle
  * des industries chimiques et connexes
