+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ccn/0044/Result_0044.js
@@ -1,13 +1,13 @@
 import React from "react";
 import MathJax from "react-mathjax-preview";
-import { ErrorBoundary } from "../../../common/ErrorBoundary";
+import { ErrorBoundary } from "../../../../common/ErrorBoundary";
 import { Container, Button } from "@cdt/ui";
 import {
   getIndemnite as getIndemniteConventionnelle,
   getSalaireRef as getSalaireRefConventionnel
 } from "./0044_indemnite";
-import { getIndemnite, getSalaireRef } from "../indemnite";
-import { SectionTitle, Highlight, SmallText } from "../stepStyles";
+import { getIndemnite, getSalaireRef } from "../../indemnite";
+import { SectionTitle, Highlight, SmallText } from "../../stepStyles";
 
 function Result_Chimie({ form }) {
   const state = form.getState();
