+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ccn/0044/__tests__/0044_indemnite.test.js
@@ -2,7 +2,7 @@ import { getIndemnite, getSalaireRef } from "../0044_indemnite";
 import {
   getIndemnite as getIndemniteLegale,
   getSalaireRef as getSalaireRefLegal
-} from "../../indemnite";
+} from "../../../indemnite";
 
 const echelon123 = [
   {
