+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ccn/0044/__tests__/Result_0044.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "react-testing-library";
 import Result_0044 from "../Result_0044";
 import { Form } from "react-final-form";
-import data from "../../__tests__/form.mock.json";
+import data from "../../../__tests__/form.mock.json";
 
 const form = {
   getState: () => data
