+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ccn/0044/index.js
@@ -1,7 +1,3 @@
-// import { Age } from "../Age";
-// import { AffiliationOpe } from "./AffilationOpe";
-// import { EchelonChimie } from "./EchelonChimie";
-// import { LicenciementEco } from "./licenciementEco";
 import Step_0044 from "./Step_0044";
 import Result_0044 from "./Result_0044";
